chore: move abstract classes to package
