+++ game/core/src/main/de/dhbw/tinf22b6/gameobject/enemy/Morglak.java
@@ -1,7 +1,6 @@
 package de.dhbw.tinf22b6.gameobject.enemy;
 
 import com.badlogic.gdx.math.Vector2;
-import de.dhbw.tinf22b6.gameobject.Enemy;
 
 public class Morglak extends Enemy {
     public Morglak(Vector2 position, int[][] rawMap) {
